update IRAS 16253 scripts

+++ IRAS16253_spectral_line_SB.py
@@ -4,22 +4,15 @@ This script was written for CASA 6.1.1/6.2
 Originally derived from DSHARP reduction scripts
 
 Datasets calibrated (in order of date observed):
-SB1: uid___A002_Xebb7f0_X690.ms
-SB2: uid___A002_Xebd1e8_X38be.ms
+SB1: 
 
-----
-LB1: uid___A002_Xf1aa15_X91aa.ms
-LB2: uid___A002_Xf1bb4a_X2738.ms
-LB3: uid___A002_Xf1bb4a_X6092.ms    
-LB4: uid___A002_Xf1bb4a_X101be.ms
-LB5: uid___A002_Xf1bb4a_X144ad.ms
+LB1: 
 
-reducer: Sacha Gavino
+reducer: 
 """
 
 """ Starting matter """
 #sys.path.append('/home/casa/contrib/AIV/science/analysis_scripts/') #CHANGE THIS TO YOUR PATH TO THE SCRIPTS!
-#sys.path.append('/users/thunter/AIV/science/analysis_scripts/')   
 import analysisUtils as au
 import analysisUtils as aU
 import string
@@ -28,7 +21,7 @@ import glob
 import numpy as np
 import sys
 import pickle
-execfile('../edisk/reduction_utils3.py', globals())
+execfile('../reduction_utils3.py', globals())
 
 
 ###############################################################
@@ -40,16 +33,17 @@ execfile('../edisk/reduction_utils3.py', globals())
 parallel=True
 
 ### if True, can run script non-interactively if later parameters properly set
-skip_plots = True	
+skip_plots = False	
 
 ### Add field names (corresponding to the field in the MS) here and prefix for 
 ### filenameing (can be different but try to keep same)
 ### Only make different if, for example, the field name has a space
-field   = {'SB':'BHR71_IRS2', 'LB':'BHR71_IRS2'}
-prefix  = 'BHR71_IRS2' 
+field   = 'IRAS16253-2429'
+prefix  = 'IRAS16253' 
 
 ### always include trailing slashes!!
-WD_path = '/lustre/cv/projects/edisk/BHR71_IRS2/'
+WD_path = '/lustre/cv/projects/edisk/IRAS16253/'
+
 SB_path = WD_path+'SB/'
 LB_path = WD_path+'LB/'
 
@@ -58,15 +52,15 @@ SB_scales = [0, 5] #[0, 5, 10, 20]
 LB_scales = [0, 5, 30]  #[0, 5, 30, 100, 200]
 
 ### automasking parameters for very extended emission
-sidelobethreshold=2.0
-noisethreshold=3.75
-lownoisethreshold=1.0
-smoothfactor=2.0
-### automasking parameters for compact emission (uncomment to use)
 #sidelobethreshold=2.0
-#noisethreshold=4.0
-#lownoisethreshold=1.5 
-#smoothfactor=1.0
+#noisethreshold=3.75
+#lownoisethreshold=1.0
+#smoothfactor=2.0
+### automasking parameters for compact emission (uncomment to use)
+sidelobethreshold=2.0
+noisethreshold=4.0
+lownoisethreshold=1.5 
+smoothfactor=1.0
 
 #read in final data_params from continuum to ensure we get the phase centers for each MS
 with open(prefix+'.pickle', 'rb') as handle:
@@ -109,15 +103,15 @@ with open(prefix+'.pickle', 'wb') as handle:
 for i in data_params.keys():
    n_tables=len(data_params[i]['selfcal_tables'])
    interp_list=['linearPD']*n_tables
-   applycal(vis=data_params[i]['vis_rescaled'].replace(WD_path,''), spw='', 
+   applycal(vis=data_params[i]['vis_rescaled'], spw='', 
          gaintable=data_params[i]['selfcal_tables'],spwmap=data_params[i]['selfcal_spwmap'], interp=interp_list, 
          calwt=True, applymode='calonly')
-   split(vis=data_params[i]['vis_rescaled'].replace(WD_path,''),outputvis=data_params[i]['vis_rescaled'].replace(WD_path,'').replace('.ms','.ms.selfcal'),datacolumn='corrected')
-   data_params[i]['vis_selfcal']=data_params[i]['vis_rescaled'].replace(WD_path,'').replace('.ms','.ms.selfcal')
+   split(vis=data_params[i]['vis_rescaled'],outputvis=data_params[i]['vis_rescaled'].replace('.ms','.ms.selfcal'),datacolumn='corrected')
+   data_params[i]['vis_selfcal']=data_params[i]['vis_rescaled'].replace('.ms','.ms.selfcal')
    ### cleanup
-   os.system('rm -rf '+data_params[i]['vis_rescaled'].replace(WD_path,''))
+   os.system('rm -rf '+data_params[i]['vis_rescaled'])
    if selectedVis=='vis_shift':
-      os.system('rm -rf '+data_params[i]['vis_shift'].replace(WD_path,''))
+      os.system('rm -rf '+data_params[i]['vis_shift'])
 
 with open(prefix+'.pickle', 'wb') as handle:
     pickle.dump(data_params, handle, protocol=pickle.HIGHEST_PROTOCOL)
@@ -130,31 +124,29 @@ with open(prefix+'.pickle', 'wb') as handle:
 ### Get channels to exclude for continuum fitting (same as the ones 
 ### we flagged for doing making continuum MS)
 for i in data_params.keys():
-   #flagchannels_string = get_flagchannels(data_params[i], prefix)
-   #print(flagchannels_string)
+   flagchannels_string = get_flagchannels(data_params[i], prefix)
+   print(flagchannels_string)
 
    ### Get spws for argument list to uvcontsub
    spws_string = get_contsub_spws_indivdual_ms(data_params[i], prefix,only_cont_spws=True)
    print(spws_string)
 
    ### Run uvcontsub on combined, self-cal applied dataset; THIS WILL TAKE MANY HOURS PER EB
-   #contsub(data_params[i]['vis_selfcal'], prefix, spw=spws_string,flagchannels=flagchannels_string,excludechans=True, combine='spw')
-   fitspw_string=fitspw_from_contdotdat(data_params[i])
+   contsub(data_params[i]['vis_selfcal'], prefix, spw=spws_string,flagchannels=flagchannels_string,excludechans=True)
    os.system('rm -rf '+prefix+'_'+i+'_spectral_line.ms')  ### remove existing spectral line MS if present
-   uvcontsub(vis=data_params[i]['vis_selfcal'],spw=spws_string,fitspw=fitspw_string,fitorder=1)
-   os.system('mv '+data_params[i]['vis_selfcal']+'.contsub '+prefix+'_'+i+'_spectral_line.ms')
+   os.system('mv '+data_params[i]['vis_selfcal'].replace('.selfcal','.selfcal.contsub')+' '+prefix+'_'+i+'_spectral_line.ms')
+   os.system('rm -rf '+data_params[i]['vis_selfcal'])
    data_params[i]['vis_contsub']=prefix+'_'+i+'_spectral_line.ms'
 
 with open(prefix+'.pickle', 'wb') as handle:
     pickle.dump(data_params, handle, protocol=pickle.HIGHEST_PROTOCOL)
-
-
 
 ###############################################################
 ################ TAR UP FINAL CONTSUBBED DATA #################
 ###############################################################
 
 for i in data_params.keys():
+   if 'SB' in i:
       os.system('rm -rf '+data_params[i]['vis_contsub']+'.tgz')
       os.system('tar czf '+data_params[i]['vis_contsub']+'.tgz '+data_params[i]['vis_contsub'])
 
@@ -171,103 +163,77 @@ for i in data_params.keys():
 
 ### Dictionary defining the spectral line imaging parameters.
 
-### generate list of MS files to image
-vislist=[]
-vislist_sb=[]
-for i in data_params.keys():
-   vislist.append(data_params[i]['vis_contsub'])
-   if 'SB' in i:
-      vislist_sb.append(data_params[i]['vis_contsub'])
-
-### Dictionary defining the spectral line imaging parameters.
-
 image_list = {
         ### C18O images
-        "C18O":dict(chanstart='-20km/s', chanwidth='0.167km/s', 
-            nchan=240, linefreq='219.56035410GHz', linespw='3',
-            robust=[0.5,2.0],imsize=4000,cellsize='0.01arcsec',uvtaper=['2000klambda']),
+        "C18O":dict(chanstart='-5.5km/s', chanwidth='0.167km/s',
+            nchan=120, linefreq='219.56035410GHz', linespw='3',
+            robust=[0.5,-0.5,0.]),
         ### 13CO images
-        "13CO":dict(chanstart='-20km/s', chanwidth='0.167km/s', 
-            nchan=240, linefreq='220.39868420GHz', linespw='1', 
-            robust=[0.5,2.0],imsize=4000,cellsize='0.01arcsec',uvtaper=['2000klambda']),
+        "13CO":dict(chanstart='-5.5km/s', chanwidth='0.167km/s',
+            nchan=120, linefreq='220.39868420GHz', linespw='1', 
+            robust=[0.5,-0.5]),
         ### 12CO images
         "12CO":dict(chanstart='-100.0km/s', chanwidth='0.635km/s', 
             nchan=315, linefreq='230.538GHz', linespw='6',
-            robust=[0.5,2.0],imsize=4000,cellsize='0.01arcsec',uvtaper=['2000klambda']),
+            robust=[0.5]),
         ### SO Images
-        "SO":dict(chanstart='-30km/s', chanwidth='0.167km/s', 
-            nchan=300, linefreq='219.94944200GHz', linespw='2',
-            robust=[0.5,2.0],imsize=4000,cellsize='0.01arcsec',uvtaper=['2000klambda']),
-        "SO_wide_binned":dict(chanstart='-42km/s', chanwidth='0.666km/s', 
-            nchan=111, linefreq='219.94944200GHz', linespw='2',
-            robust=[0.5,2.0],imsize=4000,cellsize='0.01arcsec',uvtaper=['2000klambda']),
+        "SO":dict(chanstart='-5.5km/s', chanwidth='0.167km/s', 
+            nchan=120, linefreq='219.94944200GHz', linespw='2',
+            robust=[0.5]),
         ### H2CO 3(2,1)-2(2,0) Images
-        "H2CO_3_21-2_20_218.76GHz":dict(chanstart='-20km/s', chanwidth='0.167km/s', 
-            nchan=240, linefreq='218.76006600GHz', 
-            linespw='0', robust=[0.5,2.0],imsize=4000,cellsize='0.01arcsec',uvtaper=['2000klambda']),
+        "H2CO_3_21-2_20_218.76GHz":dict(chanstart='-5.5km/s', 
+            chanwidth='0.167km/s', nchan=120, linefreq='218.76006600GHz', 
+            linespw='0', robust=[0.5]),
         ### H2CO 3(0,3)-2(0,2) Images
-        "H2CO_3_03-2_02_218.22GHz":dict(chanstart='-23km/s', chanwidth='1.34km/s', nchan=31, linefreq='218.22219200GHz', 
-            linespw='4', robust=[0.5,2.0],imsize=4000,cellsize='0.01arcsec',uvtaper=['2000klambda']),
+        "H2CO_3_03-2_02_218.22GHz":dict(chanstart='-10km/s',
+            chanwidth='1.34km/s', nchan=23, linefreq='218.22219200GHz', 
+            linespw='4', robust=[0.5]),
         ### H2CO 3(2,2)-2(2,1) Images
-        "H2CO_3_22-2_21_218.47GHz":dict(chanstart='-23km/s', chanwidth='1.34km/s', nchan=31, linefreq='218.47563200GHz',
-            linespw='4', robust=[0.5,2.0],imsize=4000,cellsize='0.01arcsec',uvtaper=['2000klambda']),
+        "H2CO_3_22-2_21_218.47GHz":dict(chanstart='-10km/s', 
+            chanwidth='1.34km/s', nchan=23, linefreq='218.47563200GHz',
+            linespw='4', robust=[0.5]),
         ### c-C3H2 217.82 GHz Images
-        "c-C3H2_217.82":dict(chanstart='-23km/s', chanwidth='1.34km/s', nchan=31, 
-             linefreq='217.82215GHz', linespw='4', robust=[0.5,2.0],imsize=4000,cellsize='0.01arcsec',uvtaper=['2000klambda']),
+        "c-C3H2_217.82":dict(chanstart='-10km/s', chanwidth='1.34km/s', 
+            nchan=23, linefreq='217.82215GHz', linespw='4', robust=[0.5]),
         ### c-C3H2 217.94 GHz Images
-        "cC3H2_217.94":dict(chanstart='-23km/s', chanwidth='1.34km/s', nchan=31, 
-             linefreq='217.94005GHz', linespw='4', robust=[0.5,2.0],imsize=4000,cellsize='0.015arcsec',uvtaper=['2000klambda']),
+        "cC3H2_217.94":dict(chanstart='-10km/s', chanwidth='1.34km/s', 
+            nchan=23, linefreq='217.94005GHz', linespw='4', robust=[0.5]),
         ### c-C3H2 218.16 GHz Images
-        "cC3H2_218.16":dict(chanstart='-23km/s', chanwidth='1.34km/s', nchan=31, 
-             linefreq='218.16044GHz', linespw='4', robust=[0.5,2.0],imsize=4000,cellsize='0.01arcsec',uvtaper=['2000klambda']),
+        "cC3H2_218.16":dict(chanstart='-10km/s', chanwidth='1.34km/s', 
+            nchan=23, linefreq='218.16044GHz', linespw='4', robust=[0.5]),
         ### DCN Images
-        "DCN":dict(chanstart='-23km/s', chanwidth='1.34km/s', nchan=31, 
-            linefreq='217.2386GHz', linespw='4', robust=[0.5,2.0],imsize=4000,cellsize='0.01arcsec',uvtaper=['2000klambda']),
+        "DCN":dict(chanstart='-10km/s', chanwidth='1.34km/s', nchan=23, 
+            linefreq='217.2386GHz', linespw='4', robust=[0.5]),
         ### CH3OH Images
-        "CH3OH":dict(chanstart='-23km/s', chanwidth='1.34km/s', nchan=31, 
-            linefreq='218.44006300GHz', linespw='4', robust=[0.5,2.0],imsize=4000,cellsize='0.01arcsec',uvtaper=['2000klambda']),
+        "CH3OH":dict(chanstart='-10km/s', chanwidth='1.34km/s', nchan=23, 
+            linefreq='218.44006300GHz', linespw='4', robust=[0.5]),
         ### SiO Images
         "SiO":dict(chanstart='-100km/s', chanwidth='1.34km/s', nchan=150, 
-            linefreq='217.10498000GHz', linespw='4', robust=[0.5,2.0],imsize=4000,cellsize='0.01arcsec',uvtaper=['2000klambda'])
+            linefreq='217.10498000GHz', linespw='4', robust=[0.5])
         }
-
-
 
 ### Loop through the spectral line images and make images.
 
-
+cellsize='0.025arcsec'
+imsize=1600
 for line in image_list:
-    print(line)
     for robust in image_list[line]["robust"]:
-        imagename = prefix+f'_SBLB_'+line+'_robust_'+str(robust)
+        imagename = prefix+f'_SB_'+line+'_robust_'+str(robust)
 
         sigma = get_sensitivity(data_params, specmode='cube', \
-                spw=[image_list[line]["linespw"]], chan=450,robust=robust,)
+                spw=[image_list[line]["linespw"]], chan=450,robust=robust,imsize=imsize,cellsize=cellsize)
 
         tclean_spectral_line_wrapper(vislist, imagename,
                 image_list[line]["chanstart"], image_list[line]["chanwidth"], 
                 image_list[line]["nchan"], image_list[line]["linefreq"], 
                 image_list[line]["linespw"], SB_scales, threshold=3.0*sigma,
-                imsize=image_list[line]["imsize"], cellsize=image_list[line]["cellsize"],
-                robust=robust, uvtaper=image_list[line]["uvtaper"],
+                imsize=imsize, cellsize=cellsize,robust=robust, 
                 sidelobethreshold=sidelobethreshold, noisethreshold=noisethreshold,
                 lownoisethreshold=lownoisethreshold,smoothfactor=smoothfactor,parallel=parallel,
                 phasecenter=data_params['SB1']['common_dir'].replace('J2000','ICRS'))
-    if selectedVis=='vis_shift':
-       tclean_spectral_line_wrapper(data_params['LB1']['vis'], imagename.replace(prefix,'temporary.pbfix'),
-        image_list[line]["chanstart"], image_list[line]["chanwidth"], 
-        image_list[line]["nchan"], image_list[line]["linefreq"], 
-        image_list[line]["linespw"][1], LB_scales, threshold=3.0*sigma,
-        imsize=image_list[line]["imsize"],
-        cellsize=image_list[line]["cellsize"],
-        robust=robust, uvtaper=image_list[line]["uvtaper"],
-        sidelobethreshold=sidelobethreshold, noisethreshold=noisethreshold,
-        lownoisethreshold=lownoisethreshold, smoothfactor=smoothfactor,
-        parallel=parallel,
-        phasecenter=data_params['SB1']['common_dir'].replace('J2000','ICRS'))
-       os.system('mv '+imagename+'.pb orig_pbimages/')
-       os.system('cp -r '+imagename.replace(prefix,'temporary.pbfix')+'.pb '+imagename+'.pb')
-       os.system('rm -rf '+imagename.replace(prefix,'temporary.pbfix')+'*')
+
+
+
 ###############################################################
 ################ CLEANUP AND FITS CONVERSION ##################
 ###############################################################
@@ -282,14 +248,9 @@ os.system("rm -rf *.pbcor* *.fits")
 
 imagelist=glob.glob('*.image') + glob.glob('*.image.tt0')
 for image in imagelist:
-    if selectedVis=='vis_shift':
-       immath(imagename=[image,image.replace('image', 'pb')],expr='IM0/IM1',outfile=image.replace('image', 'pbcor'),imagemd=image)
-    else:
-       impbcor(imagename=image, pbimage=image.replace('image', 'pb'),
-            outfile=image.replace('image', 'pbcor'))
-
-    exportfits(imagename=image.replace('image','pbcor'),fitsimage=image.replace('image','pbcor')+'.fits',overwrite=True,dropdeg=True)
-    exportfits(imagename=image,fitsimage=image+'.fits',overwrite=True,dropdeg=True)
+   impbcor(imagename=image,pbimage=image.replace('image','pb'),outfile=image.replace('image','pbcor'))
+   exportfits(imagename=image.replace('image','pbcor'),fitsimage=image.replace('image','pbcor')+'.fits',overwrite=True,dropdeg=True)
+   exportfits(imagename=image,fitsimage=image+'.fits',overwrite=True,dropdeg=True)
 
 imagelist=glob.glob('*.mask')
 for image in imagelist:
@@ -301,12 +262,14 @@ imagelist=glob.glob('*.pb') + glob.glob('*.pb.tt0')
 for image in imagelist:
    exportfits(imagename=image,fitsimage=image+'.fits',overwrite=True,dropdeg=True)
    os.system('gzip '+image+'.fits')
+
 ###############################################################
 ################# Make Plots of Everything ####################
 ###############################################################
 import sys
-sys.argv = ['../edisk/plot_final_images_SBLB.py', prefix]
-execfile('../edisk/plot_final_images_SBLB.py')
+sys.argv = ['../edisk/plot_final_images.py',prefix]
+execfile('../edisk/plot_final_images.py')
+
 
 ### Remove rescaled selfcal MSfiles
 os.system('rm -rf *rescaled.ms.*')
@@ -322,3 +285,4 @@ os.system('mv *.fits export/')
 os.system('mv *.fits.gz export/')
 os.system('mv *.tgz export/')
 os.system('mv *.pdf export/')
+
